separate our P256 helper functions
